Bump to Dalamud API 15
- Manifests + yaml + csproj SDK 14.0.1 -> 14.0.2 to API level 15
- IDutyState delegates: (object?, ushort) -> (IDutyStateEventArgs)
- ClientState.TerritoryChanged handler: ushort -> uint
- TerritoryType cast to ushort in DataCenter and RotationConfigWindow
- ImRaii: drop IEndObject/Style/Font/Color type aliases (use var),
  switch GroupDisposable checks to .Alive, popUp.Success -> .Alive,
  ChildDisposable struct pattern in ImGuiHelper member popup
- UIGlobals.PlaySoundEffect new SoundData** out args -> pass null

+++ RotationSolver.Basic/DataCenter.cs
@@ -419,7 +419,7 @@ internal static class DataCenter
 	#region Territory Info Tracking
 
 	public static Data.TerritoryInfo? Territory { get; set; }
-	public static ushort TerritoryID => Svc.ClientState.TerritoryType;
+	public static ushort TerritoryID => (ushort)Svc.ClientState.TerritoryType;
 
 	public static bool IsPvP => Territory?.IsPvP ?? false;
